requirements/dev.txt
-r common.txt
black==23.1.0
flake8==6.0.0
pep8-naming==0.13.3
flake8-quotes==3.3.2
flake8-import-order==0.18.2
django_debug_toolbar==3.8.1
parameterized==0.8.1
freezegun==1.2.2